\name{aligned_molecular_formula_annotator}
\alias{aligned_molecular_formula_annotator}
\title{
Aligned Molecular Formula Annotator
}
\description{
This function detect frequent molecular formulas across multiple samples on the aligned peak table matrix.
}
\usage{
aligned_molecular_formula_annotator(PARAM)
}
\arguments{
  \item{PARAM}{
a parameter driven from the UFA_xlsxAnalyzer module.
}
}
