\name{molecular_formula_elements_filter}
\alias{molecular_formula_elements_filter}
\title{
molecular_formula_elements_filter
}
\description{
molecular_formula_elements_filter
}
\usage{
molecular_formula_elements_filter(molecularFormulaMatrix, Elements)
}
\arguments{
  \item{molecularFormulaMatrix}{
molecularFormulaMatrix
}
  \item{Elements}{
Elements
}
}
\value{
a list of molecularFormulaMatrix and elementSorterList.
}
