\name{molecular_formulas_source_IPDB}
\alias{molecular_formulas_source_IPDB}
\title{
IPDB from a Molecular Formulas Source
}
\description{
This function produces IPDB from a molecular formulas source (A csv file).
}
\usage{
molecular_formulas_source_IPDB(PARAM_SF)
}
\arguments{
  \item{PARAM_SF}{
PARAM_SF is an internal variable of the IDSL.UFA package.
}
}
