\name{score_coefficient_evaluation}
\alias{score_coefficient_evaluation}
\title{
Score Coefficient Evaluation
}
\description{
This function evaluates the efficiency of the optimization process.
}
\usage{
score_coefficient_evaluation(PARAM_SFT)
}
\arguments{
  \item{PARAM_SFT}{
PARAM_SFT is a variable derived from the 'UFA_coefficient_xlsxAnalyzer' function
}
}
