\name{score_coefficients_optimization}
\alias{score_coefficients_optimization}
\title{
Coefficients Score Optimization
}
\description{
This function optimzes the coefficients of the score function.
}
\usage{
score_coefficients_optimization(PARAM_SFT)
}
\arguments{
  \item{PARAM_SFT}{
PARAM_SFT is a variable derived from the 'UFA_score_function_optimization_xlsxAnalyzer' function
}
}
