\name{scoreCoefficientsEvaluation}
\alias{scoreCoefficientsEvaluation}
\title{
Score Coefficient Evaluation
}
\description{
This function evaluates the efficiency of the optimization process.
}
\usage{
scoreCoefficientsEvaluation(PARAM_ScoreFunc)
}
\arguments{
  \item{PARAM_ScoreFunc}{
PARAM_ScoreFunc is a variable derived from the 'UFA_coefficient_xlsxAnalyzer' function
}
}
