\name{scoreCoefficientsOptimization}
\alias{scoreCoefficientsOptimization}
\title{
Coefficients Score Optimization
}
\description{
This function optimizes the coefficients of the score function.
}
\usage{
scoreCoefficientsOptimization(PARAM_ScoreFunc)
}
\arguments{
  \item{PARAM_ScoreFunc}{
PARAM_ScoreFunc is a variable derived from the 'UFA_score_function_optimization_xlsxAnalyzer' function
}
}
