\name{scoreCoefficientsReplicate}
\alias{scoreCoefficientsReplicate}
\title{
Zero Score Function
}
\description{
This function generates the input for the score optimization.
}
\usage{
scoreCoefficientsReplicate(PARAM_ScoreFunc)
}
\arguments{
  \item{PARAM_ScoreFunc}{
PARAM_ScoreFunc is a variable derived from the 'UFA_coefficient_xlsxAnalyzer' function
}
}
