\name{UFA_enumerated_chemical_space}
\alias{UFA_enumerated_chemical_space}
\title{
IPDBs from UFA Enumerated Chemical Space (ECS) approach
}
\description{
This function produces the isotopic profile database using the UFA enumerated chemical space (ECS) approach.
}
\usage{
UFA_enumerated_chemical_space(PARAM_MF)
}
\arguments{
  \item{PARAM_MF}{
A dataframe of the molecular formula constraints in the UFA spreadsheet
}
}
