\name{UFA_enumerated_chemical_space_xlsxAnalyzer}
\alias{UFA_enumerated_chemical_space_xlsxAnalyzer}
\title{
IPDBs from UFA Enumerated Chemical Space (ECS) xlsx Analyzer
}
\description{
This function evaluates the molecular formula generation constraints in the spreadsheet to create the isotopic profile database.
}
\usage{
UFA_enumerated_chemical_space_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
UFA spreadsheet
}
}
