\name{UFA_formula_source}
\alias{UFA_formula_source}
\title{
IPDB from a Molecular Formulas Source
}
\description{
This function produces IPDB from a molecular formula source (A .csv or .txt file).
}
\usage{
UFA_formula_source(PARAM_FormSource)
}
\arguments{
  \item{PARAM_FormSource}{
PARAM_FormSource is an internal variable of the IDSL.UFA package.
}
}
\value{
an IPDB is saved in the destination address
}
