\name{UFA_formula_source_xlsxAnalyzer}
\alias{UFA_formula_source_xlsxAnalyzer}
\title{
UFA Formula Source xlsxAnalyzer
}
\description{
This function evaluates the parameter spreadsheet for score coefficients optimization.
}
\usage{
UFA_formula_source_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
The parameter spreadsheet in the .xlsx format.
}
}
\value{
a processed parameter to feed the `UFA_molecular_formulas_source` function.
}

