\name{UFA_profile_visualizer}
\alias{UFA_profile_visualizer}
\title{
UFA Profile Visualizer
}
\description{
This function creates mass spectra comparison figures for a list of HRMS files and a vector of molecular formulas at specific retention times.
}
\usage{
UFA_profile_visualizer(PARAM_SA)
}
\arguments{
  \item{PARAM_SA}{
PARAM_SA is a variable derived from the 'UFA_profile_visualizer_xlsxAnalyzer' function.
}
}
