\name{UFA_profile_visualizer_xlsxAnalyzer}
\alias{UFA_profile_visualizer_xlsxAnalyzer}
\title{
UFA Spectra Analysis xlsxAnalyzer
}
\description{
This function processes the spreadsheet of the UFA parameters to ensure the parameter inputs are in agreement with the UFA requirements.
}
\usage{
UFA_profile_visualizer_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
UFA spreadsheet
}
}
\value{
This function returns the UFA parameters to feed the `UFA_profile_visualizer` function.
}