\name{UFA_PubChem_formula_extraction}
\alias{UFA_PubChem_formula_extraction}
\title{
UFA PubChem Formula Extraction
}
\description{
This module is to extract molecular formulas from a database.
}
\usage{
UFA_PubChem_formula_extraction(path)
}
\arguments{
  \item{path}{
path to store information
}
}
\value{
Molecular formulas in \url{https://pubchem.ncbi.nlm.nih.gov/} are stored in the provided `path` in the .txt format.
}
\references{
Kim S, Chen J, Cheng T, Gindulyte A, He J, He S, Li Q, Shoemaker BA, Thiessen PA, Yu B, Zaslavsky L, Zhang J, Bolton EE. PubChem in 2021: new data content and improved web interfaces. \emph{Nucleic Acids Res.}, 2021 Jan 8; 49(D1):D1388–D1395 , \doi{10.1093/nar/gkaa971}.
}
