\name{UFA_score_coefficient_workflow}
\alias{UFA_score_coefficient_workflow}
\title{
UFA Score Coefficient Workflow
}
\description{
This function runs the score optimization workflow.
}
\usage{
UFA_score_coefficient_workflow(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
The parameter spreadsheet in the .xlsx format.
}
}
