\name{UFA_score_function_optimization}
\alias{UFA_score_function_optimization}
\title{
UFA Score Coefficient Workflow
}
\description{
This function runs the score optimization workflow.
}
\usage{
UFA_score_function_optimization(PARAM_ScoreFunc)
}
\arguments{
  \item{PARAM_ScoreFunc}{
PARAM_ScoreFunc from the `UFA_score_function_optimization_xlsxAnalyzer` module
}
}
