\name{UFA_score_function_optimization_xlsxAnalyzer}
\alias{UFA_score_function_optimization_xlsxAnalyzer}
\title{
UFA Score Optimization xlsx Analyzer
}
\description{
This function evaluates the parameter spreadsheet for score coefficients optimization.
}
\usage{
UFA_score_function_optimization_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
The parameter spreadsheet in the .xlsx format.
}
}
