\name{UFA_workflow}
\alias{UFA_workflow}
\title{
UFA Workflow
}
\description{
This function executes the UFA workflow in order.
}
\usage{
UFA_workflow(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
UFA spreadsheet
}
}
\value{
This function orgonizes the UFA file processing for a better perofrmance using the template spreadsheet.
}
\seealso{
\url{https://ufa.idsl.me/home}
}
