\name{UFA_xlsxAnalyzer}
\alias{UFA_xlsxAnalyzer}
\title{
UFA xlsx Analyzer
}
\description{
This function processes the spreadsheet of the UFA parameters to ensure the parameter inputs are consistent with the requirements of the IDSL.UFA pipeline.
}
\usage{
UFA_xlsxAnalyzer(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
UFA spreadsheet
}
}
\value{
This function returns the UFA parameters to feed the UFA_workflow function.
}
