\name{zero_score_function}
\alias{zero_score_function}
\title{
Zero Score Function
}
\description{
This function generates the input for the score optimization.
}
\usage{
zero_score_function(PARAM_SFT)
}
\arguments{
  \item{PARAM_SFT}{
PARAM_SFT is a variable derived from the 'UFA_coefficient_xlsxAnalyzer' function
}
}
