\name{UFA_workflow}
\alias{UFA_workflow}
\title{
UFA Workflow
}
\description{
This function executes the UFA workflow in order.
}
\usage{
UFA_workflow(spreadsheet)
}
\arguments{
  \item{spreadsheet}{
UFA spreadsheet
}
}
\value{
This function organizes the UFA file processing for better performance using the template spreadsheet.
}